Shoot-on-move groundwork, Mechanism3d viz, off-axis aim, fixes
Shooter / aiming:
- ShotCalculator: add time-of-flight table + effectiveDistance() iterative
  lookahead (shoot-on-move Step 1); distance-only, drag model deferred.
- Off-axis shooter aim (6328-style): ROBOT_TO_SHOOTER transform +
  AutoAimCommand.getShooterAimHeading() = bearing + asin(lateralY/dist) +
  firingYaw, so the chassis points the SHOOTER (not the intake/forward) at
  the hub. Used by AutoAimCommand and ShootingSuperstructure.headingAtGoal.
  Placeholder transform with TODO (set from CAD).

Visualization:
- RobotMechanism3d: logs articulated component poses (hood, intake) as
  Pose3d[] for AdvantageScope; IntakerSubsystem.getPivotAngle() exposes the
  pivot angle. ascope_assets/Robot_Offseason2026 model (empty base + the two
  CAD parts as components) + config.json.

Fixes / wiring:
- IntakerConfig intakeRPS/outtakeRPS int -> double: NT returns integers as
  Long, which crashed the generated Integer wrapper on refresh().
- Bind intake to right trigger (deploy on press, retract on release).
- Standardize placeholder/tune notes into // TODO: comments.
- Update sim DS keyboard mapping.

+++ src/main/java/frc/robot/subsystems/Intaker/IntakerConfig.java
@@ -63,8 +63,10 @@ public class IntakerConfig {
         public final static double kA = 0.0;
         public final static double kS = 0.0;
 
-        public final static int intakeRPS = 0;
-        public final static int outtakeRPS = 0;
+        // NT integers come back from NetworkTables as Long, which breaks the generated
+        // Integer wrapper on refresh() — keep all tunable NT numerics as double.
+        public final static double intakeRPS = 0.0;
+        public final static double outtakeRPS = 0.0;
     }
 
     @NTParameter(tableName = "Params/"+INTAKER_PIVOT_NAME)
